requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.12
python-dotenv==1.0.0
pyjwt==2.8.0
httpx==0.27.0
browserbase
beautifulsoup4==4.12.2
playwright==1.46.0
pytest==7.4.0
